requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.1=py39hb82d6ee_1
aiosignal=1.2.0=pyhd8ed1ab_0
asttokens=2.0.5=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
blas=1.0=mkl
bleach=5.0.0=pyhd8ed1ab_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2021.10.8=h5b45459_0
catalogue=2.0.6=py39hcbf5309_2
certifi=2021.10.8=py39hcbf5309_2
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=37.0.1=py39h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
cymem=2.0.6=py39h415ef7b_3
cython=0.29.28=py39hd77b12b_0
cython-blis=0.7.7=py39h5d4886f_1
dataclasses=0.8=pyhc8e2a94_3
dataframe_image=0.1.1=py_0
debugpy=1.6.0=py39h415ef7b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
filelock=3.7.0=pyhd8ed1ab_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
frozenlist=1.3.0=py39hb82d6ee_1
gensim=4.1.2=py39hd77b12b_0
hdbscan=0.8.28=py39h5d4886f_1
huggingface_hub=0.6.0=pyhd8ed1ab_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39hcbf5309_1
importlib_metadata=4.11.3=hd8ed1ab_1
importlib_resources=5.7.1=pyhd8ed1ab_1
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.13.0=py39h832f523_0
ipython=8.3.0=py39hcbf5309_0
jedi=0.18.1=py39hcbf5309_1
jinja2=3.1.2=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h2bbff1b_0
jsonschema=4.5.1=pyhd8ed1ab_0
jupyter_client=7.3.1=pyhd8ed1ab_0
jupyter_core=4.10.0=py39hcbf5309_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
kaleido-core=0.2.1=h8ffe710_0
kiwisolver=1.3.2=py39hd77b12b_0
langcodes=3.3.0=pyhd8ed1ab_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h8d14728_1
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
llvmlite=0.38.0=py39h23ce68f_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.7=pyhd8ed1ab_0
markupsafe=2.1.1=py39hb82d6ee_1
mathjax=2.7.7=0
matplotlib=3.5.1=py39haa95532_1
matplotlib-base=3.5.1=py39hd77b12b_1
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py39hb82d6ee_1005
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
multidict=6.0.2=py39hb82d6ee_1
munkres=1.1.4=py_0
murmurhash=1.0.7=py39h415ef7b_0
nbclient=0.6.3=pyhd8ed1ab_0
nbconvert=6.5.0=pyhd8ed1ab_0
nbconvert-core=6.5.0=pyhd8ed1ab_0
nbconvert-pandoc=6.5.0=pyhd8ed1ab_0
nbformat=5.4.0=pyhd8ed1ab_0
nest-asyncio=1.5.5=pyhd8ed1ab_0
nltk=3.6.7=pyhd8ed1ab_0
numba=0.55.1=py39hb8cd55e_1
numpy=1.21.5=py39h7a0a035_2
numpy-base=1.21.5=py39hca35cd5_2
openssl=1.1.1o=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py39h2e25243_1
pandoc=2.18=h57928b3_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathy=0.6.1=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
plotly=5.6.0=pyhd3eb1b0_0
preshed=3.0.6=py39h415ef7b_2
prompt-toolkit=3.0.29=pyha770c72_0
psutil=5.9.0=py39hb82d6ee_1
pure_eval=0.2.2=pyhd8ed1ab_0
pweave=0.30.3=py_1
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=py39hb82d6ee_2
pygments=2.12.0=pyhd8ed1ab_0
pynndescent=0.5.7=pyh6c4a22f_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.9.2=py39hd77b12b_6
pyrsistent=0.18.1=py39hb82d6ee_1
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python-kaleido=0.2.1=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pywin32=303=py39hb82d6ee_0
pyyaml=6.0=py39hb82d6ee_4
pyzmq=22.3.0=py39he46f08e_2
qt=5.9.7=vc14h73c81de_0
regex=2022.4.24=py39hb82d6ee_0
requests=2.27.1=pyhd3eb1b0_0
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.0.2=py39hf11a4ad_1
scipy=1.7.3=py39h0a974cb_0
sentence-transformers=2.2.0=pyhd8ed1ab_0
sentencepiece=0.1.96=py39h2e07f2f_1
setuptools=61.2.0=py39haa95532_0
shellingham=1.4.0=pyh44b312d_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=pyhd8ed1ab_0
soupsieve=2.3.1=pyhd8ed1ab_0
spacy=3.3.0=py39h0f1f7c2_0
spacy-alignments=0.8.5=py39he5181cf_0
spacy-legacy=3.0.9=pyhd8ed1ab_0
spacy-loggers=1.0.2=pyhd8ed1ab_0
spacy-model-en_core_web_lg=3.3.0=pyhd3eb1b0_0
spacy-model-en_core_web_trf=3.3.0=pyhd3eb1b0_0
spacy-transformers=1.1.5=py39haa95532_0
sqlite=3.38.3=h2bbff1b_0
srsly=2.4.3=py39h415ef7b_1
stack_data=0.2.0=pyhd8ed1ab_0
tbb=2021.5.0=h2d74725_1
tenacity=8.0.1=py39haa95532_0
thinc=8.0.15=py39hefe7e4c_0
threadpoolctl=3.1.0=pyh8a188c0_0
tinycss2=1.1.1=pyhd8ed1ab_0
tk=8.6.11=h2bbff1b_1
tokenizers=0.12.1=py39h8c9392a_0
torchaudio=0.11.0=py39_cu113
torchvision=0.12.0=py39_cu113
tornado=6.1=py39hb82d6ee_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.2.0=pyhd8ed1ab_0
transformers=4.18.0=py39haa95532_0
typer=0.4.1=pyhd8ed1ab_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
umap-learn=0.5.3=py39hcbf5309_0
urllib3=1.26.9=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wasabi=0.9.1=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
xz=5.2.5=h8cc25b3_1
yaml=0.2.5=h8ffe710_2
yarl=1.7.2=py39hb82d6ee_2
zeromq=4.3.4=h0e60522_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h8cc25b3_2
zstd=1.4.9=h19a0ad4_0
